Move rle_encode_frame back to handler
