+++ pydicom/pixel_data_handlers/rle_handler.py
@@ -36,8 +36,8 @@ in the table below.
 """
 
 from struct import unpack
+import sys
 from typing import List, TYPE_CHECKING, cast
-
 
 try:
     import numpy as np  # type: ignore[import]
@@ -47,10 +47,11 @@ except ImportError:
 
 from pydicom.encaps import decode_data_sequence, defragment_data
 from pydicom.pixel_data_handlers.util import pixel_dtype
-from pydicom.encoders.native import rle_encode_frame  # backwards compat.
+from pydicom.encoders.native import _encode_frame
 import pydicom.uid
 
 if TYPE_CHECKING:
+    import numpy  # type: ignore[import]
     from pydicom.dataset import Dataset
 
 
@@ -388,3 +389,65 @@ def _rle_decode_segment(data: bytes) -> bytearray:
         pass
 
     return result
+
+
+# Old function kept for backwards compatibility
+def rle_encode_frame(arr: "numpy.ndarray") -> bytes:
+    """Return an :class:`numpy.ndarray` image frame as RLE encoded
+    :class:`bytearray`.
+
+    .. versionadded:: 1.3
+
+    .. deprecated:: 2.2
+
+        Use :meth:`~pydicom.dataset.Dataset.compress` instead
+
+    Parameters
+    ----------
+    arr : numpy.ndarray
+        A 2D (if *Samples Per Pixel* = 1) or 3D (if *Samples Per Pixel* = 3)
+        ndarray containing a single frame of the image to be RLE encoded.
+
+    Returns
+    -------
+    bytes
+        An RLE encoded frame, including the RLE header, following the format
+        specified by the DICOM Standard, Part 5,
+        :dcm:`Annex G<part05/chapter_G.html>`.
+    """
+    shape = arr.shape
+    if len(shape) > 3:
+        # Note: only raises if multi-sample pixel data with multiple frames
+        raise ValueError(
+            "Unable to encode multiple frames at once, please encode one "
+            "frame at a time"
+        )
+
+    # Check the expected number of segments
+    nr_segments = arr.dtype.itemsize
+    if len(shape) == 3:
+        # Number of samples * bytes per sample
+        nr_segments *= shape[-1]
+
+    if nr_segments > 15:
+        raise ValueError(
+            "Unable to encode as the DICOM standard only allows "
+            "a maximum of 15 segments in RLE encoded data"
+        )
+
+    dtype = arr.dtype
+    kwargs = {
+        'bits_allocated': arr.dtype.itemsize * 8,
+        'rows': shape[0],
+        'columns': shape[1],
+        'samples_per_pixel': 3 if len(shape) == 3 else 1,
+        'byteorder': '<',
+    }
+
+    sys_endianness = '<' if sys.byteorder == 'little' else '>'
+    byteorder = dtype.byteorder
+    byteorder = sys_endianness if byteorder == '=' else byteorder
+    if byteorder == '>':
+        arr = arr.astype(dtype.newbyteorder('<'))
+
+    return _encode_frame(arr.tobytes(), **kwargs)
